PR review ( images naming, structure, code organization)
