+++ apps/web/intl/locales/en.json
@@ -60,7 +60,7 @@
       "title": "First-hand Stories"
     },
     "campaigns": {
-      "title": "Campaigns"
+      "title": "Apply"
     },
     "news": {
       "title": "Medical news"
