+++ apps/web/intl/locales/es.json
@@ -57,7 +57,7 @@
       "title": "¿Como funciona?"
     },
     "campaigns": {
-      "title": "Campañas"
+      "title": "Postúlate"
     },
     "stories": {
       "title": "En Primera Persona"
