+++ apps/web/intl/locales/pt.json
@@ -57,7 +57,7 @@
       "title": "Como funciona?"
     },
     "campaigns": {
-      "title": "Campanhas"
+      "title": "Aplicar"
     },
     "stories": {
       "title": "Em Primeira Pessoa"
