+++ apps/web/package.json
@@ -19,7 +19,6 @@
     "ui": "*"
   },
   "devDependencies": {
-    "@headlessui/react": "^1.7.15",
     "@tailwindcss/forms": "^0.5.7",
     "@types/js-cookie": "^3.0.3",
     "@types/node": "^17.0.12",
