+++ packages/ui/components/index.ts
@@ -3,3 +3,4 @@ export * from './error';
 export * from './logo';
 export * from './share';
 export * from './social';
+export * from './video';
