+++ packages/ui/components/video/PlayVideoModal.types.ts
@@ -1,0 +1,6 @@
+export type PlayVideoProps = {
+  title?: string;
+  videoLink: string;
+  open: boolean;
+  onClose: () => void;
+};
