+++ packages/ui/components/video/index.ts
@@ -1,0 +1,1 @@
+export { PlayVideoModal } from './PlayVideoModal';
